chore(frontend): remove old journey code
closes #1651
