+++ frontend/dashboard/package.json
@@ -12,7 +12,6 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "@dagrejs/dagre": "^1.1.3",
     "@highlight-run/next": "^7.5.12",
     "@nivo/bar": "^0.87.0",
     "@nivo/core": "^0.87.0",
@@ -34,7 +33,6 @@
     "postcss": "^8.4.39",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "reactflow": "^11.11.4",
     "tailwindcss": "^3.4.4",
     "typescript": "^5.5.3"
   },
